+++ bitorch/layers/qconv1d.py
@@ -1,25 +1,36 @@
-"""Module containing the quantized convolution layer"""
+"""Module containing the quantized 1d convolution layer"""
 
 from typing import Any, Union
+
 from torch import Tensor
 from torch.nn import Conv1d, init
 from torch.nn.functional import pad, conv1d
 
-from bitorch.layers.config import config
+from bitorch import RuntimeMode
 from bitorch.quantizations import Quantization
-from bitorch.layers.qactivation import QActivation
+from .config import config
+from .extensions import DefaultImplementationMixin
+from .qactivation import QActivation
+from .qconv_mixin import QConvArgsProviderMixin
+from .register import QConv1dImplementation
 
 
 class QConv1d_NoAct(Conv1d):  # noqa: N801
-    """Quantized 1d Convolutional Layer. Has the same api as Conv1d but lets you specify a weight quantization, that is
-    applied before the convolutional operation."""
-    def __init__(self,
-                 *args: Any,
-                 weight_quantization: Union[str, Quantization] = None,
-                 pad_value: float = None,
-                 bias: bool = False,
-                 **kwargs: Any) -> None:
-        """initialization function for padding and quantization.
+    """
+    Quantized 1d Convolutional Layer. Has the same api as Conv1d but lets you specify a weight quantization, that is
+    applied before the convolutional operation.
+    """
+
+    def __init__(
+        self,
+        *args: Any,
+        weight_quantization: Union[str, Quantization] = None,
+        pad_value: float = None,
+        bias: bool = False,
+        **kwargs: Any,
+    ) -> None:
+        """
+        initialization function for padding and quantization.
 
         Args:
             weight_quantization (Union[str, Quantization], optional): quantization module or name of quantization
@@ -29,8 +40,7 @@ class QConv1d_NoAct(Conv1d):  # noqa: N801
         assert bias is False, "A QConv layer can not use a bias due to acceleration techniques during deployment."
         kwargs["bias"] = False
         super(QConv1d_NoAct, self).__init__(*args, **kwargs)
-        self._weight_quantize = config.get_quantization_function(
-            weight_quantization or config.weight_quantization)
+        self._weight_quantize = config.get_quantization_function(weight_quantization or config.weight_quantization)
         self._pad_value = pad_value or config.padding_value
 
     def _apply_padding(self, x: Tensor) -> Tensor:
@@ -64,16 +74,19 @@ class QConv1d_NoAct(Conv1d):  # noqa: N801
             stride=self.stride,
             padding=0,
             dilation=self.dilation,
-            groups=self.groups)
+            groups=self.groups,
+        )
 
 
-class QConv1d(QConv1d_NoAct):  # type: ignore
-    def __init__(self,  # type: ignore
-                 *args: Any,
-                 input_quantization: Union[str, Quantization] = None,
-                 weight_quantization: Union[str, Quantization] = None,
-                 gradient_cancellation_threshold: Union[float, None] = None,
-                 **kwargs: Any) -> None:
+class QConv1dBase(QConvArgsProviderMixin, QConv1d_NoAct):  # type: ignore
+    def __init__(
+        self,  # type: ignore
+        *args: Any,
+        input_quantization: Union[str, Quantization] = None,
+        weight_quantization: Union[str, Quantization] = None,
+        gradient_cancellation_threshold: Union[float, None] = None,
+        **kwargs: Any,
+    ) -> None:
         """initialization function for quantization of inputs and weights.
 
         Args:
@@ -84,7 +97,7 @@ class QConv1d(QConv1d_NoAct):  # type: ignore
             weight_quantization (Union[str, Quantization], optional): quantization module or name of quantization
                 function for weights. Defaults to None.
         """
-        super(QConv1d, self).__init__(*args, weight_quantization=weight_quantization, **kwargs)
+        super().__init__(*args, weight_quantization=weight_quantization, **kwargs)
         self.activation = QActivation(input_quantization, gradient_cancellation_threshold)
 
     def forward(self, input_tensor: Tensor) -> Tensor:
@@ -96,4 +109,15 @@ class QConv1d(QConv1d_NoAct):  # type: ignore
         Returns:
             Tensor: the activated and convoluted output tensor.
         """
-        return super(QConv1d, self).forward(self.activation(input_tensor))
+        return super().forward(self.activation(input_tensor))
+
+
+@QConv1dImplementation(RuntimeMode.DEFAULT)
+class QConv1d(DefaultImplementationMixin, QConv1dBase):
+    """
+    This class defines the default implementation of a QConv1d layer (which is actually implemented by QConv1dBase).
+
+    To implement a custom QConv1d implementation use QConv1dBase as a super class instead.
+    """
+
+    pass