+++ bitorch/models/base.py
@@ -4,12 +4,17 @@ from typing import Union, Type
 import torch
 from torch import nn
 
+from bitorch import RuntimeMode
 from bitorch.datasets.base import BasicDataset
-from bitorch.layers import QConv1d, QConv2d, QConv3d, QConv1d_NoAct, QConv2d_NoAct, QConv3d_NoAct
+from bitorch.layers import convert
+from bitorch.layers.qconv1d import QConv1dBase, QConv1d_NoAct
+from bitorch.layers.qconv2d import QConv2dBase, QConv2d_NoAct
+from bitorch.layers.qconv3d import QConv3dBase, QConv3d_NoAct
 
 
 class Model(nn.Module):
     """Base class for Bitorch models"""
+
     name = "None"
 
     def __init__(self, dataset: Union[BasicDataset, Type[BasicDataset]]) -> None:
@@ -53,12 +58,22 @@ class Model(nn.Module):
         for module in self._model.modules():
             if isinstance(module, (nn.Conv1d, nn.Conv2d, nn.Conv3d)):
                 # binary layers
-                if isinstance(module, (QConv1d, QConv2d, QConv3d, QConv1d_NoAct, QConv2d_NoAct, QConv3d_NoAct)):
+                if isinstance(
+                    module,
+                    (
+                        QConv1dBase,
+                        QConv2dBase,
+                        QConv3dBase,
+                        QConv1d_NoAct,
+                        QConv2d_NoAct,
+                        QConv3d_NoAct,
+                    ),
+                ):
                     nn.init.xavier_normal_(module.weight)
                 else:
                     if module.kernel_size[0] == 7:
                         # first conv layer
-                        nn.init.kaiming_normal_(module.weight, nonlinearity='relu')
+                        nn.init.kaiming_normal_(module.weight, nonlinearity="relu")
                     else:
                         # other 32-bit conv layers
                         nn.init.xavier_normal_(module.weight)
@@ -66,3 +81,6 @@ class Model(nn.Module):
                         nn.init.constant_(module.bias, 0)
             elif isinstance(module, nn.Linear):
                 nn.init.xavier_normal_(module.weight)
+
+    def convert(self, new_mode: RuntimeMode, device: torch.device = None, verbose: bool = False) -> "Model":
+        return convert(self, new_mode, device, verbose)