+++ bitorch/quantizations/base.py
@@ -5,6 +5,7 @@ import typing
 from torch import nn
 from torch.autograd.function import Function
 from typing import Any
+from warnings import warn
 
 
 class STE(Function):
@@ -13,8 +14,9 @@ class STE(Function):
     @staticmethod
     @typing.no_type_check
     def forward(
-            ctx: torch.autograd.function.BackwardCFunction,  # type: ignore
-            input_tensor: torch.Tensor) -> torch.Tensor:
+        ctx: torch.autograd.function.BackwardCFunction,  # type: ignore
+        input_tensor: torch.Tensor,
+    ) -> torch.Tensor:
         """just fowards the unchanged input_tensor.
 
         Args:
@@ -44,12 +46,19 @@ class STE(Function):
 class Quantization(nn.Module):
     """superclass for quantization modules"""
 
-    name = "None"
-    bitwidth = -1
+    name: str = "None"
+    bit_width: int = -1
+
+    @property
+    def bitwidth(self) -> int:
+        warn("Attribute 'bitwidth' is deprecated, use 'bit_width' instead.", DeprecationWarning, stacklevel=2)
+        return self.bit_width
 
     def quantize(self, x: torch.Tensor) -> torch.Tensor:
-        """quantize the input tensor. It is recommended to use a torch.Function to also maniputlate backward behaiviour. See
-        the implementations of sign or dorefa quantization functions for more examples.
+        """Quantize the input tensor.
+
+        It is recommended to use a torch.Function to also manipulate backwards behavior.
+        See the implementations of sign or dorefa quantization functions for more examples.
 
         Args:
             x (torch.Tensor): the input to be quantized