+++ bitorch/quantizations/dorefa.py
@@ -16,7 +16,7 @@ class WeightDoReFa(Quantization):
     """
 
     name = "weightdorefa"
-    bitwidth = config.dorefa_bits
+    bit_width = config.dorefa_bits
 
     def __init__(self, bits: Union[int, None] = None) -> None:
         """Initiates quantization bits.
@@ -25,8 +25,8 @@ class WeightDoReFa(Quantization):
             bits (int, optional): number of bits to quantize into. Defaults to None.
         """
         super(WeightDoReFa, self).__init__()
-        self.bitwidth = bits or config.dorefa_bits
-        self._max_value = 2 ** self.bitwidth - 1
+        self.bit_width = bits or config.dorefa_bits
+        self._max_value = 2**self.bit_width - 1
 
     def quantize(self, x: torch.Tensor) -> torch.Tensor:
         """DoReFas the tensor to desired bit resolution using weight dorefa.
@@ -47,8 +47,8 @@ class InputDoReFaFunction(Function):
     @staticmethod
     @typing.no_type_check
     def forward(
-            ctx: torch.autograd.function.BackwardCFunction,  # type: ignore
-            input_tensor: torch.Tensor, bits: int) -> torch.Tensor:
+        ctx: torch.autograd.function.BackwardCFunction, input_tensor: torch.Tensor, bits: int  # type: ignore
+    ) -> torch.Tensor:
         """quantizes input tensor and forwards it.
 
         Args:
@@ -59,7 +59,7 @@ class InputDoReFaFunction(Function):
         Returns:
             torch.Tensor: the quantized input tensor
         """
-        max_value = 2 ** bits - 1
+        max_value = 2**bits - 1
 
         quantized_tensor = torch.round(torch.clamp(input_tensor, 0, 1) * max_value) / max_value
         return quantized_tensor
@@ -87,7 +87,7 @@ class InputDoReFa(Quantization):
     """
 
     name = "inputdorefa"
-    bitwidth = config.dorefa_bits
+    bit_width = config.dorefa_bits
 
     def __init__(self, bits: Union[int, None] = None) -> None:
         """Initiates quantization bits.
@@ -96,7 +96,7 @@ class InputDoReFa(Quantization):
             bits (int, optional): number of bits to quantize into. Defaults to None.
         """
         super(InputDoReFa, self).__init__()
-        self.bitwidth = bits or config.dorefa_bits
+        self.bit_width = bits or config.dorefa_bits
 
     def quantize(self, x: torch.Tensor) -> torch.Tensor:
         """DoReFas the tensor to desired bit resolution.
@@ -107,5 +107,4 @@ class InputDoReFa(Quantization):
         Returns:
             torch.Tensor: DoReFaed tensor x
         """
-
-        return InputDoReFaFunction.apply(x, self.bitwidth)
+        return InputDoReFaFunction.apply(x, self.bit_width)