+++ bitorch/quantizations/sign.py
@@ -1,8 +1,9 @@
 """Sign Function Implementation"""
-from typing import Tuple, Union, Optional
+
+import typing
 
 import torch
-import typing
+
 from .base import Quantization, STE
 
 
@@ -10,8 +11,9 @@ class SignFunction(STE):
     @staticmethod
     @typing.no_type_check
     def forward(
-            ctx: torch.autograd.function.BackwardCFunction,  # type: ignore
-            input_tensor: torch.Tensor) -> torch.Tensor:
+        ctx: torch.autograd.function.BackwardCFunction,  # type: ignore
+        input_tensor: torch.Tensor,
+    ) -> torch.Tensor:
         """Binarize the input tensor using the sign function
 
         Args:
@@ -22,8 +24,7 @@ class SignFunction(STE):
             torch.Tensor: the sign tensor
         """
         sign_tensor = torch.sign(input_tensor)
-        sign_tensor = torch.where(sign_tensor == 0, torch.tensor(
-            1., device=sign_tensor.device), sign_tensor)
+        sign_tensor = torch.where(sign_tensor == 0, torch.tensor(1.0, device=sign_tensor.device), sign_tensor)
         return sign_tensor
 
 
@@ -31,7 +32,7 @@ class Sign(Quantization):
     """Module for applying the sign function with straight through estimator in backward pass"""
 
     name = "sign"
-    bitwidth = 1
+    bit_width = 1
 
     def quantize(self, x: torch.Tensor) -> torch.Tensor:
         """Forwards the tensor through the sign function.