+++ examples/pytorch_lightning/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for image classification example."""